Add support for reading/writing arguments from/to .netconfig
The [dotnetconfig](https://dotnetconfig.org) format allows hierarchical
tool settings to be persisted and read in a standards way. By adding
this support to the tool, all options can now be persisted to the
`.netconfig` in the current directory to streamline subsequent runs.

When running `dotnet serve .... --save`, the options will be saved
to the current directory `.netconfig` file and can be inspected to
learn the format. Afterwards, the options can be moved if desired
to another `.netconfig` file in any ancestor directory so it's reused
across runs from any descendent folder. Or it can also be saved to the
global or system locations too.

Updated the readme too to showcase this new feature.

As an example, moved a few arguments from the launchSettings.json to
a .netconfig in the project directory, which now augments the startup
arguments from config :).

In order to add tests that exercise the new fallbacks and options
augmentation, modified the Program.cs a bit so it can be instantiated
and its RunAsync overriden so that tests don't actually run but
can exercise the full options parsing and overriding/defaulting from
this PR.