Add support for reading/writing arguments from/to .netconfig (#64)
The [dotnetconfig](https://dotnetconfig.org) format allows hierarchical
tool settings to be persisted and read in a standards way. By adding
this support to the tool, all options can now be persisted to the
`.netconfig` in the current directory to streamline subsequent runs.

When running `dotnet serve .... --save`, the options will be saved
to the current directory `.netconfig` file and can be inspected to
learn the format. Afterwards, the options can be moved if desired
to another `.netconfig` file in any ancestor directory so it's reused
across runs from any descendent folder. Or it can also be saved to the
global or system locations too.